requirements.txt
black==21.9b0
brotlipy==0.7.0
certifi==2021.5.30
cffi
charset-normalizer
click==8.0.2
cryptography
Django==2.2.24
djangorestframework==3.9.4
idna
importlib-metadata==4.8.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycparser
pyOpenSSL
PySocks
pytz==2021.3
regex==2021.10.8
requests
six
sqlparse==0.4.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
pandas==1.3.5
